+++ Player.java
@@ -44,19 +44,19 @@ public class Player {
 
         if (p instanceof Sunflower) {
             if (currentTick - Sunflower.getLastPlacedSunflower() < p.getPlacementCooldown()) {
-                int remaining = p.getPlacementCooldown() - (currentTick - Sunflower.getLastPlacedSunflower());
+                double remaining = p.getPlacementCooldown() - (currentTick - Sunflower.getLastPlacedSunflower());
                 System.out.println("Sunflower still in cooldown for placement. Wait for " + remaining + " turns.");
                 return false;
             }
         } else if (p instanceof Peashooter) {
             if (currentTick - Peashooter.getLastPlacedPeashooter() < p.getPlacementCooldown()) {
-                int remaining = p.getPlacementCooldown() - (currentTick - Peashooter.getLastPlacedPeashooter());
+                double remaining = p.getPlacementCooldown() - (currentTick - Peashooter.getLastPlacedPeashooter());
                 System.out.println("Peashooter still in cooldown for placement. Wait for " + remaining + " turns.");
                 return false;
             }
         } else if (p instanceof CherryBomb) {
             if (currentTick - CherryBomb.getLastPlacedCherry() < p.getPlacementCooldown()) {
-                int remaining = p.getPlacementCooldown() - (currentTick - CherryBomb.getLastPlacedCherry());
+                double remaining = p.getPlacementCooldown() - (currentTick - CherryBomb.getLastPlacedCherry());
                 System.out.println("Cherry still in cooldown for placement. Wait for " + remaining + " turns.");
                 return false;
             }
